feat: Separate `UserInterface` and `ReportManagement` services.
