+++ src/main/java/org/spin/grpc/service/LogsInfo.java
@@ -38,7 +38,6 @@ import org.compiere.util.CLogger;
 import org.compiere.util.Env;
 import org.compiere.util.Util;
 import org.compiere.wf.MWFProcess;
-import org.spin.backend.grpc.common.ChatEntry;
 import org.spin.backend.grpc.common.ProcessLog;
 import org.spin.backend.grpc.logs.EntityChat;
 import org.spin.backend.grpc.logs.EntityChat.ConfidentialType;
@@ -60,6 +59,7 @@ import org.spin.backend.grpc.logs.ListUserActivitesResponse;
 import org.spin.backend.grpc.logs.ListWorkflowLogsRequest;
 import org.spin.backend.grpc.logs.ListWorkflowLogsResponse;
 import org.spin.backend.grpc.logs.LogsGrpc.LogsImplBase;
+import org.spin.backend.grpc.user_interface.ChatEntry;
 import org.spin.backend.grpc.logs.RecentItem;
 import org.spin.backend.grpc.wf.WorkflowProcess;
 import org.spin.base.db.LimitUtil;
