+++ src/main/proto/import_file_loader.proto
@@ -18,10 +18,10 @@ option java_multiple_files = true;
 option java_package = "org.spin.backend.grpc.form.import_file_loader";
 option java_outer_classname = "ADempiereImportFileLoader";
 
-import "base_data_type.proto";
-import "business.proto";
-import "google/protobuf/struct.proto";
 import "google/api/annotations.proto";
+import "google/protobuf/struct.proto";
+
+import "base_data_type.proto";
 
 // Base URL
 // /import-loader/
