+++ src/main/proto/payment_allocation.proto
@@ -18,10 +18,10 @@ option java_multiple_files = true;
 option java_package = "org.spin.backend.grpc.form.payment_allocation";
 option java_outer_classname = "ADempierePaymentAllocation";
 
-import "google/protobuf/struct.proto";
-import "business.proto";
-import "google/protobuf/timestamp.proto";
 import "google/api/annotations.proto";
+import "google/protobuf/timestamp.proto";
+
+import "base_data_type.proto";
 
 // Base URL
 // /payment-allocation/
