+++ src/main/proto/payment_print_export.proto
@@ -1,5 +1,5 @@
 /************************************************************************************
- * Copyright (C) 2018-2023 E.R.P. Consultores y Asociados, C.A.                     *
+ * Copyright (C) 2018-present E.R.P. Consultores y Asociados, C.A.                  *
  * Contributor(s): Edwin Betancourt EdwinBetanc0urt@outlook.com                     *
  * This program is free software: you can redistribute it and/or modify             *
  * it under the terms of the GNU General Public License as published by             *
@@ -18,11 +18,10 @@ option java_multiple_files = true;
 option java_package = "org.spin.backend.grpc.payment_print_export";
 option java_outer_classname = "ADempierePaymentPrintExport";
 
-import "base_data_type.proto";
-import "business.proto";
-import "core_functionality.proto";
-import "google/protobuf/struct.proto";
 import "google/api/annotations.proto";
+
+import "base_data_type.proto";
+import "core_functionality.proto";
 
 // Base URL
 // /payment-print/
