+++ src/main/proto/time_record.proto
@@ -1,5 +1,5 @@
 /************************************************************************************
- * Copyright (C) 2012-2023 E.R.P. Consultores y Asociados, C.A.                     *
+ * Copyright (C) 2018-present E.R.P. Consultores y Asociados, C.A.                  *
  * Contributor(s): Edwin Betancourt EdwinBetanc0urt@outlook.com                     *
  * This program is free software: you can redistribute it and/or modify             *
  * it under the terms of the GNU General Public License as published by             *
@@ -19,10 +19,11 @@ option java_multiple_files = true;
 option java_package = "org.spin.backend.grpc.form.time_record";
 option java_outer_classname = "ADempiereTimeRecord";
 
-import "core_functionality.proto";
+import "google/api/annotations.proto";
 import "google/protobuf/struct.proto";
 import "google/protobuf/timestamp.proto";
-import "google/api/annotations.proto";
+
+import "core_functionality.proto";
 
 // Base URL
 // /time-record/
