+++ src/main/proto/user_customization.proto
@@ -19,8 +19,8 @@ option java_multiple_files = true;
 option java_package = "org.spin.backend.grpc.user_customization";
 option java_outer_classname = "ADempiereUserCustomization";
 
-import "google/protobuf/empty.proto";
 import "google/api/annotations.proto";
+import "google/protobuf/empty.proto";
 
 // Base URL
 // /customizations/
