+++ src/main/proto/workflow.proto
@@ -1,5 +1,5 @@
 /************************************************************************************
- * Copyright (C) 2018-2023 E.R.P. Consultores y Asociados, C.A.                     *
+ * Copyright (C) 2018-present E.R.P. Consultores y Asociados, C.A.                  *
  * Contributor(s): Edwin Betancourt EdwinBetanc0urt@outlook.com                     *
  * This program is free software: you can redistribute it and/or modify             *
  * it under the terms of the GNU General Public License as published by             *
@@ -19,10 +19,11 @@ option java_multiple_files = true;
 option java_package = "org.spin.backend.grpc.wf";
 option java_outer_classname = "ADempiereWorkflow";
 
-import "base_data_type.proto";
+import "google/api/annotations.proto";
 import "google/protobuf/empty.proto";
 import "google/protobuf/timestamp.proto";
-import "google/api/annotations.proto";
+
+import "base_data_type.proto";
 
 // Base URL
 // /workflow/
